+++ src/OpenTelemetry.Exporter.OpenTelemetryProtocol/OtlpExporterOptions.cs
@@ -19,6 +19,8 @@ using System.Diagnostics;
 using System.Net.Http;
 #endif
 using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Options;
 using OpenTelemetry.Internal;
 using OpenTelemetry.Metrics;
 using OpenTelemetry.Trace;
@@ -189,5 +191,13 @@ namespace OpenTelemetry.Exporter
         /// Gets a value indicating whether <see cref="Endpoint" /> was modified via its setter.
         /// </summary>
         internal bool ProgrammaticallyModifiedEndpoint { get; private set; }
+
+        internal static void RegisterOtlpExporterOptionsFactory(IServiceCollection services)
+        {
+            services.RegisterOptionsFactory(
+                (sp, configuration, name) => new OtlpExporterOptions(
+                    configuration,
+                    sp.GetRequiredService<IOptionsMonitor<BatchExportActivityProcessorOptions>>().Get(name)));
+        }
     }
 }
